% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{fire_sf}
\alias{fire_sf}
\title{sf object with fire spots from 2015 to 2022 of the Brazilian Amazon}
\format{
A tibble with 4428447 rows and 16 variables:
datahora: Date time of observation,
satelite: Name of the satellite,
pais: Country name,
estado: State name,
municipio: Town name,
bioma: Biome,
diasemchuv: Number of days without rain,
precipitac: TODO,
riscofogo: Fire risk,
latitude: Latitude,
longitude: Longitude,
frp: Fire Radiative Power,
xy_id: Subarea id (based on their centroid coordinates),
prodes_code: Land Use-Cover code in PRODES 2021,
prodes_date: View date in PRODES 2021,
year: PRODES year,
geom: geometry field,
}
\usage{
data(fire_sf)
}
\description{
An sf object containing points corresponding to the
data from satellite NP-375 downloaded from INPE's queimadas platform.
The dataset contains data from 2015-08-01 to 2022-07-26.
}
\keyword{datasets}
