% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{subarea_sf}
\alias{subarea_sf}
\title{sf object with the subareas of deforestation alerts from 2008 to 2021}
\format{
A tibble with 807069 rows and 2 variables:
subarea_id: ID of the subarea,
geom: geometry field.
}
\usage{
data(subarea_sf)
}
\description{
An sf object containing polygons corresponding to the dataset
subarea_tb.
}
\keyword{datasets}
