% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{subarea_tb}
\alias{subarea_tb}
\title{Data Frame of the subareas of deforestation alerts from 2008 to 2021}
\format{
A tibble with 1085482 rows and 17 variables:
CLASSNAME: Name of warning or deforestation,
QUADRANT: Out of use. CBERS AWFI quadrant,
PATH_ROW: Path and row of the used satellite images,
VIEW_DATE: Date of the images used to identify a warning,
SENSOR: Name of the sensor which take the images,
SATELLITE: Name of the satellite which took the images,
AREAUCKM: Warning area intesecting a conservation unit,
UC: Name of conservation unit,
AREAMUNKM: Warning area intersecting a municipality,
MUNICIPALI:Name of municipality intersecting a warning,
GEOCODIBGE: IBGE's code for the municipality,
UF: Brazilian state name,
xy_id: Subarea ID based on its centroid coordinates,
subarea_ha: Subarea extent in hectares,
in_prodes: Does the subarea intersect the PRODES mask,
year: PRODES year,
data_source: Source of the data; either DETER or PRODES,
}
\usage{
data(subarea_tb)
}
\description{
A dataset containing a tibble object with no-spatial data of
subareas of DETER's deforestation alerts and PRODES's deforestation areas.
DETER subareas are segments of DETER alerts which have continuity over
time, that is, their shape do not change over time. Capitalized names are
the DETER attributes corresponding to the warnings. The code required to
download the data required and create these dataset are available in the
directory data-raw. When PRODES is updated (currently PRODES 2021
v20220915 is used), also updated its classes codes (see file
PDigital2000_2021_AMZ_raster_v20220915_bioma.txt) in the script
04_process_deter.R at the definition fo the variable prodes_classes.
}
\keyword{datasets}
